+++ src/items/rays/mod.rs
@@ -11,8 +11,8 @@ pub const ID: u8 = 2;
 
 struct Arrow {
     angle: f32, //0.0 = 0, 0.25 = 45, 0.5 = 90 and so on.
-    x: i32,
-    y: i32,
+    x: f32,
+    y: f32,
 }
 
 struct ButtonState {
@@ -32,8 +32,8 @@ pub fn start(display: &mut BoxDisplay, sdl_context: &mut SdlContext, write: &Wri
 
     let mut arrow = Arrow {
         angle: 0.0,
-        x: (window_width / 2) as i32,
-        y: (window_height / 2) as i32,
+        x: (window_width / 2)  as f32,
+        y: (window_height / 2) as f32,
     };
 
     let ray_color: Color = Color::RGB(228, 133, 230);
@@ -50,15 +50,15 @@ pub fn start(display: &mut BoxDisplay, sdl_context: &mut SdlContext, write: &Wri
 
     let mut modifiers: Vec<Modifier> = Vec::new();
 
-    let mut modifiers_label: Label = Label::new(0, 0, 8, write, None);
+    let mut modifiers_label: Label = Label::new(0.0, 0.0, 8, write, None);
 
     modifiers.push(Modifier {
         name: String::from("Precision"),
         slider: Slider::new(
             0,                            //Minimum
             200,                          //Max
-            80,                           //X
-            (window_height - 40) as i32,  //Y
+            80.0,                           //X
+            (window_height - 40) as f32,  //Y
             120,                          //Length
             SliderType::SliderHorizontal, //Type
         ),
@@ -68,8 +68,8 @@ pub fn start(display: &mut BoxDisplay, sdl_context: &mut SdlContext, write: &Wri
         slider: Slider::new(
             10,
             1000,
-            80,
-            (window_height - 80) as i32,
+            80.0,
+            (window_height - 80) as f32,
             120,
             SliderType::SliderHorizontal,
         ),
@@ -79,8 +79,8 @@ pub fn start(display: &mut BoxDisplay, sdl_context: &mut SdlContext, write: &Wri
         slider: Slider::new(
             0,
             100,
-            80,
-            (window_height - 120) as i32,
+            80.0,
+            (window_height - 120) as f32,
             120,
             SliderType::SliderHorizontal,
         ),
@@ -181,7 +181,7 @@ pub fn start(display: &mut BoxDisplay, sdl_context: &mut SdlContext, write: &Wri
         arrow.turn(&right_arrow);
         let mut next_angle: f32 = arrow.angle;
 
-        let (mut pos_end_x, mut pos_end_y): (i32, i32) = ray_corner(
+        let (mut pos_end_x, mut pos_end_y): (f32, f32) = ray_corner(
             (arrow.x, arrow.y),
             (window_width, window_height),
             &mut next_angle,
@@ -198,7 +198,7 @@ pub fn start(display: &mut BoxDisplay, sdl_context: &mut SdlContext, write: &Wri
         display.canvas.set_draw_color(Color::RGB(26, 90, 186));
 
         for _ in 0..ray_cycles {
-            let (temp_x, temp_y): (i32, i32) = ray_corner(
+            let (temp_x, temp_y): (f32, f32) = ray_corner(
                 (pos_end_x, pos_end_y),
                 (window_width, window_height),
                 &mut next_angle,
@@ -213,11 +213,11 @@ pub fn start(display: &mut BoxDisplay, sdl_context: &mut SdlContext, write: &Wri
         }
 
         for modifier in modifiers.iter() {
-            modifiers_label.set_pos(modifier.slider.x + 80, modifier.slider.y - 4);
+            modifiers_label.set_pos(modifier.slider.x + 80.0, modifier.slider.y - 4.0);
             modifiers_label.update_text(modifier.slider.get_value_ref().to_string());
             let _ = modifiers_label.draw(display);
 
-            modifiers_label.set_pos(modifier.slider.x, modifier.slider.y - 22);
+            modifiers_label.set_pos(modifier.slider.x, modifier.slider.y - 22.0);
             modifiers_label.update_text(String::from(&modifier.name));
             let _ = modifiers_label.draw(display);
             if modifier
@@ -236,38 +236,38 @@ pub fn start(display: &mut BoxDisplay, sdl_context: &mut SdlContext, write: &Wri
 }
 
 //return a position based on the angle.
-fn angle_pos(x: i32, y: i32, mut angle: f32, distance: f32) -> (i32, i32) {
+fn angle_pos(x: f32, y: f32, mut angle: f32, distance: f32) -> (f32, f32) {
     angle *= PI;
     (
-        x + (distance * angle.sin()) as i32,
-        y + (distance * angle.cos()) as i32,
+        x + (distance * angle.sin()),
+        y + (distance * angle.cos()),
     )
 }
 
-fn rounder(value: &mut i32, max: u32) {
-    if *value > max as i32 {
-        *value = max as i32
-    } else if *value < 0i32 {
-        *value = 0i32
+fn rounder(value: &mut f32, max: u32) {
+    if *value > max as f32 {
+        *value = max as f32
+    } else if *value < 0f32 {
+        *value = 0f32
     }
 }
 
 //This function simulates a ray and returns the possible collision position.
 //Along with changing the angle inserted.
 fn ray_corner(
-    (x, y): (i32, i32),
+    (x, y): (f32, f32),
     (width, height): (u32, u32),
     angle: &mut f32,
     precision: u32,
     ray_length: f32,
-) -> (i32, i32) {
+) -> (f32, f32) {
     for i in 1..(precision + 1) {
         let (mut px, mut py) = angle_pos(x, y, *angle, (ray_length / precision as f32) * i as f32);
-        if px > width as i32 || px < 0 {
+        if px > width as f32 || px < 0.0 {
             rounder(&mut px, width);
             *angle *= -1.0;
             return (px, py);
-        } else if py > height as i32 || py < 0 {
+        } else if py > height as f32 || py < 0.0 {
             rounder(&mut py, height);
             *angle = *angle * -1.0 + 1.0;
             return (px, py);
